% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{IC}
\alias{IC}
\title{Intervalo de Confianza}
\usage{
IC(sigma, n, media)
}
\arguments{
\item{sigma}{Sigma de la poblacion}

\item{n}{muestra}

\item{media}{promedio de la muestra}
}
\description{
Intervalo de Confianza
}
